Move resources, update classpath and cleanup
Move resource files from src/java/resources to res and update .classpath to use src as the source folder. Remove old decompiled java files under src/java and restore/clean sources under src (package refactor). Perform small code cleanups: remove unused imports and unnecessary casts in ImgFix, use diamond operator for ArrayList, adjust import ordering/spacing, and other minor refactors across enemies, helpz, inputs, managers, scenes and UI classes.

+++ TDGame/src/towers/Tower.java
@@ -54,7 +54,7 @@ public class Tower {
     }
 
     public boolean isCooldownOver() {
-        return (float)this.cdTick >= this.cooldown;
+        return this.cdTick >= this.cooldown;
     }
 
     public void resteCooldown() {
